+++ CHANGELOG.md
@@ -62,6 +62,7 @@ released by then.
 
 ### Fixed
 
+- Domain name letter case mismatches in DNS rewrites ([#3351]).
 - Conflicts between IPv4 and IPv6 DNS rewrites ([#3343]).
 - Letter case mismatches in `CNAME` filtering ([#3335]).
 - Occasional breakages on network errors with DNS-over-HTTP upstreams ([#3217]).
@@ -105,6 +106,7 @@ released by then.
 [#3257]: https://github.com/AdguardTeam/AdGuardHome/issues/3257
 [#3335]: https://github.com/AdguardTeam/AdGuardHome/issues/3335
 [#3343]: https://github.com/AdguardTeam/AdGuardHome/issues/3343
+[#3351]: https://github.com/AdguardTeam/AdGuardHome/issues/3351